Move ProductArea index and fetch into service
