+++ web/app/components/inputs/product-select/index.hbs
@@ -11,7 +11,7 @@
         @isSaving={{@isSaving}}
         @renderOut={{@renderOut}}
         @icon={{this.icon}}
-        class="w-80 product-select-dropdown-list"
+        class="product-select-dropdown-list w-80"
         ...attributes
       >
         <:item as |dd|>
@@ -32,7 +32,7 @@
         @placement={{@placement}}
         @isSaving={{@isSaving}}
         @renderOut={{@renderOut}}
-        class="w-[300px] product-select-dropdown-list"
+        class="product-select-dropdown-list w-[300px]"
         ...attributes
       >
         <:anchor as |dd|>
@@ -68,12 +68,12 @@
         </:item>
       </X::DropdownList>
     {{/if}}
-  {{else if this.fetchProducts.isRunning}}
+  {{else if this.productAreas.fetch.isRunning}}
     <FlightIcon data-test-product-select-spinner @name="loading" />
   {{else}}
     <div
       class="absolute top-0 left-0"
-      {{did-insert (perform this.fetchProducts)}}
+      {{did-insert (perform this.productAreas.fetch)}}
     ></div>
   {{/if}}
 </div>
